+++ skrl/models/torch/gaussian.py
@@ -77,23 +77,23 @@ class GaussianMixin:
               )
             )
         """
-        self._clip_actions = clip_actions and isinstance(self.action_space, gymnasium.Space)
+        self._g_clip_actions = clip_actions and isinstance(self.action_space, gymnasium.Space)
 
-        if self._clip_actions:
-            self._clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
-            self._clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
+        if self._g_clip_actions:
+            self._g_clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
+            self._g_clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
 
-        self._clip_log_std = clip_log_std
-        self._log_std_min = min_log_std
-        self._log_std_max = max_log_std
+        self._g_clip_log_std = clip_log_std
+        self._g_log_std_min = min_log_std
+        self._g_log_std_max = max_log_std
 
-        self._log_std = None
-        self._num_samples = None
-        self._distribution = None
+        self._g_log_std = None
+        self._g_num_samples = None
+        self._g_distribution = None
 
         if reduction not in ["mean", "sum", "prod", "none"]:
             raise ValueError("reduction must be one of 'mean', 'sum', 'prod' or 'none'")
-        self._reduction = (
+        self._g_reduction = (
             torch.mean
             if reduction == "mean"
             else torch.sum if reduction == "sum" else torch.prod if reduction == "prod" else None
@@ -129,26 +129,26 @@ class GaussianMixin:
         mean_actions, log_std, outputs = self.compute(inputs, role)
 
         # clamp log standard deviations
-        if self._clip_log_std:
-            log_std = torch.clamp(log_std, self._log_std_min, self._log_std_max)
+        if self._g_clip_log_std:
+            log_std = torch.clamp(log_std, self._g_log_std_min, self._g_log_std_max)
 
-        self._log_std = log_std
-        self._num_samples = mean_actions.shape[0]
+        self._g_log_std = log_std
+        self._g_num_samples = mean_actions.shape[0]
 
         # distribution
-        self._distribution = Normal(mean_actions, log_std.exp())
+        self._g_distribution = Normal(mean_actions, log_std.exp())
 
         # sample using the reparameterization trick
-        actions = self._distribution.rsample()
+        actions = self._g_distribution.rsample()
 
         # clip actions
-        if self._clip_actions:
-            actions = torch.clamp(actions, min=self._clip_actions_min, max=self._clip_actions_max)
+        if self._g_clip_actions:
+            actions = torch.clamp(actions, min=self._g_clip_actions_min, max=self._g_clip_actions_max)
 
         # log of the probability density function
-        log_prob = self._distribution.log_prob(inputs.get("taken_actions", actions))
-        if self._reduction is not None:
-            log_prob = self._reduction(log_prob, dim=-1)
+        log_prob = self._g_distribution.log_prob(inputs.get("taken_actions", actions))
+        if self._g_reduction is not None:
+            log_prob = self._g_reduction(log_prob, dim=-1)
         if log_prob.dim() != actions.dim():
             log_prob = log_prob.unsqueeze(-1)
 
@@ -169,9 +169,9 @@ class GaussianMixin:
             >>> print(entropy.shape)
             torch.Size([4096, 8])
         """
-        if self._distribution is None:
+        if self._g_distribution is None:
             return torch.tensor(0.0, device=self.device)
-        return self._distribution.entropy().to(self.device)
+        return self._g_distribution.entropy().to(self.device)
 
     def get_log_std(self, role: str = "") -> torch.Tensor:
         """Return the log standard deviation of the model
@@ -187,7 +187,7 @@ class GaussianMixin:
             >>> print(log_std.shape)
             torch.Size([4096, 8])
         """
-        return self._log_std.repeat(self._num_samples, 1)
+        return self._g_log_std.repeat(self._g_num_samples, 1)
 
     def distribution(self, role: str = "") -> torch.distributions.Normal:
         """Get the current distribution of the model
@@ -203,4 +203,4 @@ class GaussianMixin:
             >>> print(distribution)
             Normal(loc: torch.Size([4096, 8]), scale: torch.Size([4096, 8]))
         """
-        return self._distribution
+        return self._g_distribution